rename to DictionaryPoints
