+++ packages/site/src/lib/components/home/DictionaryPoints.svelte
